Implement basic typing functionality and text switching
This introduces the ability to type a piece of text and have it display
which segments are completed, erroneous, or untyped.

This also implements text switching, so you can seamlessly move between
typing the main text and a side text by just starting to type the side
text and continuing to type until only one matching side text remains.
Note that this means side texts *need* to be prefixed with a character
that won't be used at all in the main text. Otherwise, we could get
kidnapped by the side text if we try to type that character in the main
text and become soft locked.