Reimplement input manager with better backspace functionality
Backspace can now return you from a side-text to the main text or from a
side-text into the selection state.

Additionally, this change implements a more robust function for
determining the current state of the input manager (MAIN_TEXT,
SELECTION, or SIDE_TEXT) which will trigger the debugger if some part of
the state is wrong or out of sync.