Implement CursorText UI component
This component renders text with a cursor that can be moved between
characters. It comes with a bunch of utility functions for setting the
text, wrapping the text to fit within the UI, updating an attached
TextState, etc.

It can link to the InputManager and TextStates so the UI can update
automatically as the input_state and text progress changes.

Two new signals were added to TextState with this change: updated and
newline.