Implement spellbook with spells that modify the main text
Spells are added as SpellText nodes in the Spellbook scene. The actual
implementation of their callback functions is in spellbook.gd.

The spellbook can be dropped into any scene where it has access to the
InputManager and a CursorText containing the main text. It will animate
when focus is moved to/away from the side texts.

Implemented: shorten text, remove punctuation, remove capital letters.

Negative effects are not implemented.