Add ControlView3D scene to more easily render UI in 3D space
This scene can be dropped into any other 3D scene. A child Control node
can then be added under the ControlView3D, which will render in the
viewport, both in-editor and in-game. Note that only one Control node
can be added as a direct child, so if multiple Control nodes are needed,
they should be nested under a top-level Control node.

The size can be changed by just modifying the ControlView3D's scale
property.