Implement text-locking functionality
Calling the lock() function on a TextState will set the state to locked.
When in the locked state, it will not allow modifications to or resets
of the text progress. A signal, lock_changed, is emitted when the locked
state changes.

The InputManager is also updated to ignore locked texts when choosing
candidates as well as to re-sync TextStates when they become unlocked.

The CursorText UI component will re-render when the locked state of its
linked TextState changes. Right now, it shows the text as blocks when
the text is locked because a typo was made.