Ensure start menu only plays typing sound while still present
As one types "Sign the Contract" an issue arises where we still try to play the typing sound after the start menu node has been removed from the scene tree in favor of the main scene. This results in the following error:

````
Playback can only happen when a node is inside the scene tree
```

Thus, we can simply check that the start menu is still present before playing the sound.