Improve scene loading speed with async loading
Now loading of the intro and main scenes start asynchronously from the
_ready function of the title menu, and loading of the main scene starts
asynchronously from the _ready function of the intro scene. This
prevents main thread freezes when loading the next scene.

Note that the main scene, in an ideal world, would only be preloaded
from the intro scene (the scene that switches to it), but since web
builds don't support threading, the async loading doesn't actually seem
to be async. So we just load both the intro and main scenes as soon as
the title menu scene starts. This means the title menu doesn't actually
appear until both the preloaded scenes load, but this delay is hidden
since the previous screen is the game loading screen.